refactored database schemas as the file was becoming too long
