+++ code/server/database/models/schematics/index.js
@@ -1,9 +1,25 @@
-const general = require('./schemas/generalSchemas.js')
-const subDocs = require('./schemas/subDocs.js')
 const discriminators = require('./schemas/discriminators.js')
+const jummah = require('./schemas/jummah.js')
+const location = require('./schemas/location.js')
+const institution = require('./schemas/institution.js')
+const timing = require('./schemas/timing.js')
+const user = require('./schemas/user.js')
+const announcement = require('./schemas/announcement.js')
+const setting = require('./schemas/setting.js')
+const notification = require('./schemas/notification.js')
+const shortenedURL = require('./schemas/shortenedURL.js')
+const verificationCode = require('./schemas/verificationCode.js')
 
 module.exports = {
-    ...general,
-    ...subDocs,
-    ...discriminators
+    jummah,
+    location,
+    institution,
+    timing,
+    user,
+    announcement,
+    setting,
+    notification,
+    shortenedURL,
+    verificationCode,
+    discriminators
 }
