+++ code/server/database/models/schematics/schemas/institution.js
@@ -1,0 +1,37 @@
+const mongoose = require('mongoose')
+
+const institution = new mongoose.Schema({
+    name: {
+        type: String,
+        required: true,
+        minlength: 1
+    },
+    abbreviatedName: {
+        type: String,
+        required: false,
+        default: '__NO-ABBREVIATED-NAME__'
+    },
+    timezone: {
+        type: String,
+        required: true,
+        minlength: 1
+    },
+    confirmed: {
+        type: Boolean,
+        required: false,
+        default: false
+    },
+    country: {
+        type: String,
+        required: true,
+        minlength: 1
+    },
+    state: {
+        type: String,
+        required: false,
+        minlength: 1
+    }
+},
+{ timestamps: true })
+
+module.exports = institution
